Update Excess manangement behaviour
Up to now and as delta to the original pv_excess script by inventocasa: Using the minimum value from the array of PV excess power logs from the last minute (or whatever you min switch time is) to device about turning on, off and adjustment of current for dynamic appliances. The original code use the arithmetic mean (see original comments why thing this is not the best idea).

With this change I use the minímum for the toggle on, increasing current decisions and the median for turning off or reducing current.

Reason: Be conservative on increasing / turning on. Also charging BEVs increase the current mostly slowly. To avoid miscalculations if the current has not yet been realized by the car, minimum is a good choice.
For turning off / reducing we do not want to overreat to small spikes in the PV production, especially as  lot of installations have batteries meanwhile which can bridge short reductions in PV power. The median is a better choice than arithmetic mean for such analysis.